+++ dbms/postgresql/src/main/java/is/codion/dbms/postgresql/PostgreSQLDatabase.java
@@ -125,11 +125,6 @@ final class PostgreSQLDatabase extends AbstractDatabase {
 	}
 
 	@Override
-	public boolean subqueryRequiresAlias() {
-		return true;
-	}
-
-	@Override
 	public int maximumNumberOfParameters() {
 		return MAXIMUM_STATEMENT_PARAMETERS;
 	}