+++ changelog.md
@@ -5,6 +5,8 @@ Codion Change Log
 ### is.codion.common.core
 - Value.set() and map() return value removed.
 - AbstractValue now only compares the values when set if the notification policy requires it.
+### is.codion.common.db
+- Database.subqueryRequiresAlias() removed, related refactoring.
 ### is.codion.common.model
 - DefaultConditionModel.DefaultSetCondition refactored.
 - MultiSelection.Indexes now extends Value.